PIL-1291 Importing wallet leads to frozen screen and does not successfully import (#2980)
* shown the permission screen

* welcome back screen changes
