+++ src/components/Modal/Modal.js
@@ -74,6 +74,7 @@ type Props = {|
   hideModalContentWhileAnimating?: boolean,
   style?: any,
   backdropDismissable?: boolean,
+  isSwipeClose?: boolean,
 |};
 
 type State = {|
@@ -148,7 +149,7 @@ class Modal extends React.Component<Props, State> {
   render() {
     // swipeDirection is passed explicitly to work around issues with using
     // spread operator on union types
-    const { children, swipeDirection, backdropDismissable = false, ...props } = this.props;
+    const { children, swipeDirection, backdropDismissable = false, isSwipeClose = false, ...props } = this.props;
     const { isVisible } = this.state;
 
     // render contents direrctly with overlay-like background if the modal was
@@ -164,7 +165,7 @@ class Modal extends React.Component<Props, State> {
         isVisible={isVisible}
         propagateSwipe
         swipeDirection={swipeDirection ?? 'down'}
-        onSwipeComplete={this.close}
+        onSwipeComplete={!isSwipeClose ? this.close : null}
         onBackdropPress={!backdropDismissable ? this.close : null}
         onBackButtonPress={this.close}
         avoidKeyboard
