+++ src/constants/navigationConstants.js
@@ -67,7 +67,7 @@ export const CHANGE_PIN_CONFIRM_NEW_PIN = 'CHANGE_PIN_CONFIRM_NEW_PIN';
 export const WELCOME = 'WELCOME';
 export const ONBOARDING_FLOW = 'ONBOARDING_FLOW';
 export const NEW_WALLET = 'NEW_WALLET';
-export const NEW_PROFILE = 'NEW_PROFILE';
+export const WELCOME_BACK = 'WELCOME_BACK';
 export const IMPORT_WALLET = 'IMPORT_WALLET';
 export const IMPORT_WALLET_LEGALS = 'IMPORT_WALLET_LEGALS';
 export const SET_WALLET_PIN_CODE = 'SET_WALLET_PIN_CODE';
