--- src/screens/NewProfile/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "NewProfile.js",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./NewProfile.js"
-}
