+++ src/screens/WelcomeBack/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "WelcomeBack.js",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./WelcomeBack.js"
+}
